Feat: UserHandler 및 UserService findById의 userId를 String 타입으로 변경

+++ src/main/kotlin/com/youquiz/user/handler/UserHandler.kt
@@ -19,7 +19,7 @@ class UserHandler(
 
     suspend fun findById(request: ServerRequest): ServerResponse =
         request.pathVariable("id").let {
-            ServerResponse.ok().bodyValueAndAwait(userService.findById(it.toLong()))
+            ServerResponse.ok().bodyValueAndAwait(userService.findById(it))
         }
 
     suspend fun findByUsername(request: ServerRequest): ServerResponse =
